backend/requirements.txt
fastapi==0.100.0  
uvicorn==0.34.0
pytest==7.4.0
httpx==0.28.1
openai==1.12.0
python-decouple==3.8